RavenDB-24991 Cleanup and small fixes
