+++ src/Raven.Studio/typescript/components/shell/chatbot/Chatbot.stories.tsx
@@ -63,6 +63,18 @@ export const Endpoints: StoryObj<typeof Chatbot> = {
         commonInit();
 
         const { chatbot } = mockStore;
+
+        for (let i = 0; i < 10; i++) {
+            const id = `endpoint-${_.uniqueId()}`;
+            chatbot.with_attachedContextAdded({
+                id,
+                type: "Endpoints Responses",
+                label: id,
+                value: '{"Name":"Auto/Categories/ById()","SourceType":"Documents","Type":"AutoMap"',
+                state: "included",
+            });
+        }
+
         chatbot.with_messages(ChatbotStubs.messagesWithEndpoints());
 
         return <ChatbotInLayout />;
